+++ Userland/SampleCodeModule/shell/testing/test_util.c
@@ -69,7 +69,7 @@ void endless_loop_print(uint64_t wait) {
   int64_t pid = getpid();
 
   while (1) {
-    printf("%d ", pid);
+    printf("%d ", (int)pid);
     bussy_wait(wait);
   }
 }
